+++ src/ragwell/_generated/api/retrieval/search_v1_projects_project_id_search_post.py
@@ -59,6 +59,11 @@ def _parse_response(
 
         return response_404
 
+    if response.status_code == 409:
+        response_409 = ErrorResponse.from_dict(response.json())
+
+        return response_409
+
     if response.status_code == 422:
         response_422 = ErrorResponse.from_dict(response.json())
 
@@ -89,6 +94,11 @@ def _parse_response(
         response_500 = ErrorResponse.from_dict(response.json())
 
         return response_500
+
+    if response.status_code == 502:
+        response_502 = ErrorResponse.from_dict(response.json())
+
+        return response_502
 
     if response.status_code == 503:
         response_503 = ErrorResponse.from_dict(response.json())